+++ data/tasks.txt
@@ -1,1 +1,1 @@
-Todo`hihi`N/endDeadline`test2`N`2023-01-23 23:09/endEvent`null`N`2023-12-08`2024-12-12/endDeadline`test3`N`2023-09-12 18:09/end
+duke.task.Todo`hihi`N/endDeadline`test2`N`2023-01-23 23:09/endEvent`null`N`2023-12-08`2024-12-12/endDeadline`test3`N`2023-09-12 18:09/end
