+++ src/main/java/duke/command/AddCommand.java
@@ -1,3 +1,15 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.Parser;
+import duke.Storage;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.Todo;
+import duke.TaskList;
+import duke.Ui;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
